Moves create bill run flow to separate controller.  Adds stricter URL param validation in flow
