+++ src/internal/modules/billing/controllers/bill-run.js
@@ -1,15 +1,9 @@
 'use strict';
 
-const uuid = require('uuid/v4');
-const { selectBillingTypeForm, billingTypeFormSchema } = require('../forms/billing-type');
-const { selectBillingRegionForm, billingRegionFormSchema } = require('../forms/billing-region');
 const confirmForm = require('../forms/confirm-form');
 const { cancelOrConfirmBatchForm } = require('../forms/cancel-or-confirm-batch');
 const services = require('internal/lib/connectors/services');
-const forms = require('shared/lib/forms');
-const { get, kebabCase, groupBy, sortBy } = require('lodash');
-const queryString = require('querystring');
-const helpers = require('@envage/water-abstraction-helpers');
+const { groupBy, sortBy } = require('lodash');
 const batchService = require('../services/batch-service');
 const transactionsCSV = require('../services/transactions-csv');
 const csv = require('internal/lib/csv-download');
@@ -17,163 +11,8 @@ const { logger } = require('internal/logger');
 const mappers = require('../lib/mappers');
 const titleCase = require('title-case');
 const { pluralize } = require('shared/lib/pluralize');
-const urlJoin = require('url-join');
 const moment = require('moment');
 const Boom = require('@hapi/boom');
-const routing = require('../lib/routing');
-
-const { TWO_PART_TARIFF } = require('../lib/bill-run-types');
-const seasons = require('../lib/seasons');
-
-const getSessionForm = (request) => {
-  return request.yar.get(get(request, 'query.form'));
-};
-
-const clearSessionForm = (request) => {
-  request.yar.clear(get(request, 'query.form'));
-};
-
-const getBillingRegions = async () => {
-  const { data } = await services.water.regions.getRegions();
-  return data;
-};
-
-const getRegionUrl = (selectedBillingType, selectedTwoPartTariffSeason, formKey) => {
-  const path = urlJoin(
-    '/billing/batch/region',
-    kebabCase(selectedBillingType),
-    kebabCase(selectedTwoPartTariffSeason)
-  );
-
-  return formKey
-    ? `${path}?${queryString.stringify({ form: formKey })}`
-    : path;
-};
-
-const getBatchDetails = (request, billingRegionForm) => {
-  const {
-    selectedBillingType,
-    selectedBillingRegion,
-    selectedTwoPartTariffSeason
-  } = forms.getValues(billingRegionForm);
-
-  const financialYear = helpers.charging.getFinancialYear();
-  const financialYearEnding = selectedTwoPartTariffSeason === seasons.WINTER_AND_ALL_YEAR
-    ? financialYear - 1
-    : financialYear;
-
-  const batch = {
-    userEmail: request.defra.user.user_name,
-    regionId: selectedBillingRegion,
-    batchType: selectedBillingType,
-    financialYearEnding,
-    isSummer: selectedTwoPartTariffSeason === seasons.SUMMER
-  };
-  return batch;
-};
-
-/**
- * Step 1a of create billing batch flow - display form to select type
- * i.e. Annual, Supplementary, Two-Part Tariff
- * @param {*} request
- * @param {*} h
- */
-const getBillingBatchType = async (request, h) => {
-  const sessionForm = getSessionForm(request);
-  if (sessionForm) { clearSessionForm(request); }
-
-  return h.view('nunjucks/form', {
-    ...request.view,
-    back: '/manage',
-    form: sessionForm || selectBillingTypeForm(request)
-  });
-};
-
-/**
- * Step 1b - receive posted step 1a data
- * @param {*} request
- * @param {*} h
- */
-const postBillingBatchType = async (request, h) => {
-  const billingTypeForm = forms.handleRequest(selectBillingTypeForm(request), request, billingTypeFormSchema(request));
-
-  if (billingTypeForm.isValid) {
-    const { selectedBillingType, twoPartTariffSeason } = forms.getValues(billingTypeForm);
-    return h.redirect(getRegionUrl(
-      selectedBillingType,
-      selectedBillingType === TWO_PART_TARIFF ? twoPartTariffSeason : ''
-    ));
-  }
-
-  const key = uuid();
-  request.yar.set(key, billingTypeForm);
-  return h.redirect('/billing/batch/type?' + queryString.stringify({ form: key }));
-};
-
-/**
- * Step 2a - display selelct region form
- * @param {*} request
- * @param {*} h
- */
-const getBillingBatchRegion = async (request, h) => {
-  const sessionForm = getSessionForm(request);
-  if (sessionForm) { clearSessionForm(request); }
-
-  const regions = await getBillingRegions();
-
-  return h.view('nunjucks/form', {
-    ...request.view,
-    back: '/billing/batch/type',
-    form: sessionForm || selectBillingRegionForm(request, regions)
-  });
-};
-
-/**
- * Step 2b received step 2a posted data
- * try to create a new billing run batch
- * redirect to waiting page
- * @param {*} request
- * @param {*} h
- */
-const postBillingBatchRegion = async (request, h) => {
-  const regions = await getBillingRegions();
-  const billingRegionForm = forms.handleRequest(selectBillingRegionForm(request, regions), request, billingRegionFormSchema);
-
-  if (!billingRegionForm.isValid) {
-    const { selectedBillingType, selectedTwoPartTariffSeason } = forms.getValues(billingRegionForm);
-
-    const key = uuid();
-    request.yar.set(key, billingRegionForm);
-
-    return h.redirect(getRegionUrl(selectedBillingType, selectedTwoPartTariffSeason, key));
-  }
-
-  try {
-    const batch = getBatchDetails(request, billingRegionForm);
-    const { data } = await services.water.billingBatches.createBillingBatch(batch);
-    const path = routing.getBillingBatchRoute(data.batch, false);
-    return h.redirect(path);
-  } catch (err) {
-    if (err.statusCode === 409) {
-      return h.redirect(`/billing/batch/${err.error.existingBatch.id}/exists`);
-    }
-    throw err;
-  }
-};
-
-/**
- * If the Bill run for the type and region exists then display a basic summary page
- * @param {*} request
- * @param {*} h
- */
-const getBillingBatchExists = async (request, h) => {
-  return h.view('nunjucks/billing/batch-exist', {
-    ...request.view,
-    today: new Date(),
-    back: '/billing/batch/region',
-    batch: request.pre.batch
-  });
-};
 
 const mapInvoice = invoice => ({
   ...invoice,
@@ -365,14 +204,7 @@ const getBillingBatchEmpty = async (request, h) => {
 
 exports.getBillingBatchList = getBillingBatchList;
 exports.getBillingBatchSummary = getBillingBatchSummary;
-exports.getBillingBatchExists = getBillingBatchExists;
 exports.getBillingBatchInvoice = getBillingBatchInvoice;
-
-exports.getBillingBatchType = getBillingBatchType;
-exports.postBillingBatchType = postBillingBatchType;
-
-exports.getBillingBatchRegion = getBillingBatchRegion;
-exports.postBillingBatchRegion = postBillingBatchRegion;
 
 exports.getBillingBatchCancel = getBillingBatchCancel;
 exports.postBillingBatchCancel = postBillingBatchCancel;
