+++ src/internal/modules/billing/forms/billing-region.js
@@ -42,15 +42,17 @@ const selectBillingRegionForm = (request, regions) => {
   return f;
 };
 
-const billingRegionFormSchema = {
+const getRegionIds = regions => regions.map(region => region.regionId);
+
+const billingRegionFormSchema = regions => ({
   csrf_token: Joi.string().uuid().required(),
-  selectedBillingRegion: Joi.string().uuid().required(),
+  selectedBillingRegion: Joi.string().uuid().required().valid(getRegionIds(regions)),
   selectedBillingType: Joi.string().required(),
   selectedTwoPartTariffSeason: Joi.string().allow('').when('selectedBillingType', {
     is: TWO_PART_TARIFF,
     then: Joi.string().required()
   })
-};
+});
 
 exports.selectBillingRegionForm = selectBillingRegionForm;
 exports.billingRegionFormSchema = billingRegionFormSchema;
