+++ src/internal/modules/billing/routes/bill-run.js
@@ -8,79 +8,6 @@ const isAcceptanceTestTarget = ['local', 'dev', 'development', 'test', 'preprod'
 const preHandlers = require('../pre-handlers');
 if (isAcceptanceTestTarget) {
   module.exports = {
-    getBillingBatchType: {
-      method: 'GET',
-      path: '/billing/batch/type',
-      handler: controller.getBillingBatchType,
-      config: {
-        auth: { scope: allowedScopes },
-        description: 'select bill run type',
-        plugins: {
-          viewContext: {
-            pageTitle: 'Which kind of bill run do you want to create?',
-            activeNavLink: 'notifications'
-          }
-        }
-      }
-    },
-
-    postBillingBatchType: {
-      method: 'POST',
-      path: '/billing/batch/type',
-      handler: controller.postBillingBatchType,
-      config: {
-        auth: { scope: allowedScopes },
-        description: 'post handler for select type'
-      }
-    },
-
-    getBillingBatchRegion: {
-      method: 'GET',
-      path: '/billing/batch/region/{billingType}/{season?}',
-      handler: controller.getBillingBatchRegion,
-      config: {
-        auth: { scope: allowedScopes },
-        description: 'select bill run region',
-        plugins: {
-          viewContext: {
-            pageTitle: 'Select the region',
-            activeNavLink: 'notifications'
-          }
-        }
-      }
-    },
-
-    postBillingBatchRegion: {
-      method: 'POST',
-      path: '/billing/batch/region',
-      handler: controller.postBillingBatchRegion,
-      config: {
-        auth: { scope: allowedScopes },
-        description: 'post handler for receiving the selected region'
-      }
-    },
-
-    getBillingBatchExists: {
-      method: 'GET',
-      path: '/billing/batch/{batchId}/exists',
-      handler: controller.getBillingBatchExists,
-      config: {
-        pre: [{ method: preHandlers.loadBatch, assign: 'batch' }],
-        auth: { scope: allowedScopes },
-        description: 'If a bill run exist, warn user and display short summary',
-        plugins: {
-          viewContext: {
-            pageTitle: 'A bill run already exists',
-            activeNavLink: 'notifications'
-          }
-        },
-        validate: {
-          params: {
-            batchId: Joi.string().uuid()
-          }
-        }
-      }
-    },
 
     getBillingBatchSummary: {
       method: 'GET',
